feat: :sparkles: refactor session command imports for improved structure and consistency
